Reference dataset used for testing and demonstration (anndata)

A reference dataset for accute myeloid leukemia was included in this package for demonstration and testing. The data was originally published in Triana et al. 2021. The SCUBA authors downsampled the original Seurat object to use in the package for automated testing, and converted it into other object formats. The cell types provided by Triana et al. were also condensed into 10 generalized cell types to facilitate demonstration of SCUBA visualization capabilities. Details on the operations performed from the original object are provided in this script in the SCUBA manuscript repository.

Usage

AML_h5ad()

Source

The dataset was obtained from the Figshare. For more information on the operations performed on the original object, see the SCUBA manuscript repository.

Details

AML_h5ad() is called as a function, with parentheses. This is because the dataset is in Python, and can't be loaded using the typical process of loading data included with R packages.

Examples

# These examples require a functional Python installation with prerequisite
# packages installed to work
# Please see our website for more details
# https://amc-heme.github.io/SCUBA/index.html#installation
# 
# The examples may take a while (about 10 seconds) to run the first 
# time they are executed in a session due to the time required to 
# initialize a Python environment. 
# R Studio does not display a spinner while these run, so the "run_examples"
# link may not appear to do anything until the examples are finished. 
AML_h5ad()
#> AnnData object with n_obs × n_vars = 250 × 462
#>     obs: 'nCount_RNA', 'nFeature_RNA', 'nCount_AB', 'nFeature_AB', 'nCount_BOTH', 'nFeature_BOTH', 'BOTH_snn_res.0.9', 'seurat_clusters', 'Prediction_Ind', 'BOTH_snn_res.1', 'ClusterID', 'Batch', 'x', 'y', 'x_mean', 'y_mean', 'cor', 'ct', 'prop', 'meandist', 'cDC', 'B.cells', 'Myelocytes', 'Erythroid', 'Megakaryocte', 'Ident', 'RNA_snn_res.0.4', 'condensed_cell_type'
#>     var: 'vst.mean', 'vst.variance', 'vst.variance.expected', 'vst.variance.standardized'
#>     obsm: 'X_pca', 'X_umap', 'protein'
#>     layers: 'counts'

# Summary of metadata variables in object
meta_varnames(AML_h5ad())
#>  [1] "nCount_RNA"          "nFeature_RNA"        "nCount_AB"          
#>  [4] "nFeature_AB"         "nCount_BOTH"         "nFeature_BOTH"      
#>  [7] "BOTH_snn_res.0.9"    "seurat_clusters"     "Prediction_Ind"     
#> [10] "BOTH_snn_res.1"      "ClusterID"           "Batch"              
#> [13] "x"                   "y"                   "x_mean"             
#> [16] "y_mean"              "cor"                 "ct"                 
#> [19] "prop"                "meandist"            "cDC"                
#> [22] "B.cells"             "Myelocytes"          "Erythroid"          
#> [25] "Megakaryocte"        "Ident"               "RNA_snn_res.0.4"    
#> [28] "condensed_cell_type"